+++ Sanctuary.Census.ClientData/Abstractions/Services/IClientDataCacheService.cs
@@ -12,7 +12,7 @@ public interface IClientDataCacheService : IDataCacheService
     /// <summary>
     /// Gets the cached <see cref="AreaDefinition"/> objects.
     /// </summary>
-    IReadOnlyDictionary<AssetZone, List<AreaDefinition>>? AreaDefinitions { get; }
+    IReadOnlyDictionary<AssetZone, IReadOnlyList<AreaDefinition>>? AreaDefinitions { get; }
 
     /// <summary>
     /// Gets the cached <see cref="ClientItemDatasheetData"/> objects.